backend/requirements.txt
# =============================================================================
# Future Prediction AI - Python Dependencies
# =============================================================================

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
websockets>=12.0

# Anthropic (Claude API)
anthropic==0.39.0

# Data Processing
pandas==2.2.2
numpy>=1.26.2
pydantic==2.9.2
pydantic-settings==2.5.2

# Market Data
yfinance>=0.2.44
# pandas-ta removed: numpy conflict with langchain-chroma; fallback calculations used instead

# HTTP & Async
httpx==0.27.2
aiohttp==3.10.5
requests==2.32.3

# Web Scraping
beautifulsoup4==4.12.3
lxml==5.3.0
feedparser==6.0.11

# Reddit API
praw==7.7.1

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2

# Caching
redis==5.0.8
cachetools==5.5.0

# NLP & Sentiment
textblob==0.18.0
vaderSentiment==3.3.2

# Date & Time
python-dateutil==2.9.0
pytz==2024.1

# Utilities
python-dotenv==1.0.1
tenacity==9.0.0
structlog==24.4.0

# Google Trends
pytrends==4.9.2

# Development & Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2

# Type checking & Linting
mypy==1.11.2
ruff==0.6.8

# FRED Economic Data
fredapi==0.5.2

# RAG Pipeline
chromadb>=0.5.0
langchain-chroma>=0.2.0,<0.3.0
langchain-huggingface>=0.1.0,<0.2.0
langchain-text-splitters>=0.3.0,<0.4.0
langchain-core>=0.3.15,<0.4.0
langchain-community>=0.3.0,<0.4.0
sentence-transformers>=3.0.0

# Prompt Versioning
pyyaml==6.0.2
